Requisito 2 tudo funcionando

+++ WebApplication1/Controllers/PessoaModelsController.cs
@@ -59,12 +59,40 @@ namespace WebApplication1.Controllers
         {
             //if (ModelState.IsValid)
             //{
-            pessoaModel.Situação = true;//n sei se ta certo
+            pessoaModel.Situação = true;
             _context.Add(pessoaModel);
+
+            //Parte de checagem de dados
+            if (pessoaModel.QuantidadeFilhos < 0)
+            {
+                ModelState.AddModelError("", "A quantidade de filhos tem que ser maior ou igual a zero!");
+                return View(pessoaModel);
+            }
+            if (pessoaModel.Salario < 1200 || pessoaModel.Salario > 13000)
+            {
+                ModelState.AddModelError("", "O salário tem que ser entre 1200 e 13000!");
+                return View(pessoaModel);
+            }
+
+            var verifyEmail = await _context.PessoaModel.FirstOrDefaultAsync(m => m.Email == pessoaModel.Email);
+
+            if (verifyEmail != null && verifyEmail.Codigo != pessoaModel.Codigo)
+            {
+                ModelState.AddModelError("", "Este email já está cadastrado!");
+                return View(pessoaModel);
+            }
+
+            DateTime defaultDate = new DateTime(1990, 1, 1,00,00,00);
+            if (pessoaModel.DataNascimento < defaultDate)
+            {
+                ModelState.AddModelError("", "A data de nascimento tem que ser depois de 01/01/1990!");
+                return View(pessoaModel);
+            }
+
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
             //}
-            return View(pessoaModel);
+            //return View(pessoaModel);
         }
 
         // GET: PessoaModels/Edit/5
@@ -96,6 +124,41 @@ namespace WebApplication1.Controllers
                 return NotFound();
             }
 
+            //Parte de checagem de dados
+            if (pessoaModel.Situação == false)
+            {
+                ModelState.AddModelError("", "Não é possível editar uma pessoa com a situação Inativa");
+                return View(pessoaModel);
+            }
+
+            if (pessoaModel.QuantidadeFilhos < 0)
+            {
+                ModelState.AddModelError("", "A quantidade de filhos tem que ser maior ou igual a zero!");
+                return View(pessoaModel);
+            }
+
+            if (pessoaModel.Salario < 1200 || pessoaModel.Salario > 13000)
+            {
+                ModelState.AddModelError("", "O salário tem que ser entre 1200 e 13000!");
+                return View(pessoaModel);
+            }
+
+            var verifyEmail =  _context.PessoaModel.Where(m => m.Email.Equals(pessoaModel.Email) && m.Codigo != pessoaModel.Codigo);
+
+            if (verifyEmail.Count() > 0)
+            {
+                ModelState.AddModelError("", "Este email já está cadastrado!");
+                return View(pessoaModel);
+            }
+
+            DateTime defaultDate = new DateTime(1990, 1, 1, 00, 00, 00);
+            if (pessoaModel.DataNascimento < defaultDate)
+            {
+                ModelState.AddModelError("", "A data de nascimento tem que ser depois de 01/01/1990!");
+                return View(pessoaModel);
+            }
+
+            //Editando
             if (ModelState.IsValid)
             {
                 try
@@ -130,7 +193,6 @@ namespace WebApplication1.Controllers
             var pessoaModel = await _context.PessoaModel
                 .FirstOrDefaultAsync(m => m.Codigo == id);
 
-
             if (pessoaModel == null)
             {
                 return NotFound();
@@ -145,6 +207,18 @@ namespace WebApplication1.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var pessoaModel = await _context.PessoaModel.FindAsync(id);
+            if (pessoaModel == null)
+            {
+                return NotFound();
+            }
+
+            //Parte de checagem de dados
+            if (pessoaModel.Situação == true)
+            {
+                ModelState.AddModelError("", "Não é possível excluir uma pessoa com a situação Ativa");
+                return View(pessoaModel);
+            }
+
             _context.PessoaModel.Remove(pessoaModel);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -195,17 +269,19 @@ namespace WebApplication1.Controllers
             {
                 _context.Update(pessoaModel);
                 await _context.SaveChangesAsync();
-            } catch (DbUpdateConcurrencyException)
+            }
+            catch (DbUpdateConcurrencyException)
             {
                 if (!PessoaModelExists(pessoaModel.Codigo))
                 {
                     return NotFound();
-                } else
+                }
+                else
                 {
                     throw;
                 }
             }
-            
+
             return RedirectToAction(nameof(Index));
         }
